src/MiNET/MiNET/Plugins/Command.cs
using System.Collections.Generic;
using Newtonsoft.Json;

namespace MiNET.Plugins
{
	public class Commands : Dictionary<string, Command>
	{
	}

	public class Command
	{
		public Version[] Versions { get; set; }
	}

	public class Version
	{
		[JsonProperty(propertyName: "version")]
		public int CommandVersion { get; set; }

		public string[] Aliases { get; set; }
		public string Description { get; set; }
		public string Permission { get; set; }
		public bool RequiresChatPerms { get; set; }

		[JsonProperty(propertyName: "requires_edu")]
		public bool RequiresEdu { get; set; }

		public Dictionary<string, Overload> Overloads { get; set; }
	}


	public class Overload
	{
		public Input Input { get; set; }
		public Output Output { get; set; }
		public Parser Parser { get; set; }
	}

	public class Input
	{
		public Parameter[] Parameters { get; set; }
	}

	public class Output
	{
		[JsonProperty(propertyName: "format_strings")]
		public string[] FormatStrings { get; set; }

		public Parameter[] Parameters { get; set; }
	}

	public class Parser
	{
		public string Tokens { get; set; }
	}

	public class Parameter
	{
		public string Name { get; set; }
		public string Type { get; set; }

		[JsonProperty(propertyName: "enum_type")]
		public string EnumType { get; set; }

		[JsonProperty(propertyName: "enum_values")]
		public string[] WnumValues { get; set; }

		public bool Optional { get; set; }
	}
}